+++ src/app/talkroom/question/page.tsx
@@ -1,0 +1,5 @@
+const page = () => {
+  return <div>question</div>;
+};
+
+export default page;
